Refactored Piwik\Access into Piwik\Authorization\Authorization without singleton
Piwik\Access is kept for BC and is now a proxy to the new class
